// cypress/DataParameters/Enums/approverUser.ts
export enum ApproverUser {
    elisaUrban = "select2-approver_user_id-result-lpj6-2",
    felipePugliesi = "select2-approver_user_id-result-5yqt-3",
    camilaTeixeira = "select2-approver_user_id-result-dm2r-4",
    andreRoedel = "select2-approver_user_id-result-ljz6-5",
    julianaRavanello = "select2-approver_user_id-result-lg2i-6",
    anaMariaSantana = "select2-approver_user_id-result-7z8a-7",
    brunoUrban = "select2-approver_user_id-result-f9x9-8",
    thaisRighetto = "select2-approver_user_id-result-6mqf-9",
    vitorUrban = "select2-approver_user_id-result-lqqm-10",
    denizeVergilinoDaSilva = "select2-approver_user_id-result-bnp5-11",
    fernandoRodrigues = "select2-approver_user_id-result-x9k4-12",
    cirieleBoeiraCataneo = "select2-approver_user_id-result-iw86-13",
    lucianoSaraceni = "select2-approver_user_id-result-rzqy-14",
    catianeMoraes = "select2-approver_user_id-result-27gd-15",
    silvanaUemura = "select2-approver_user_id-result-f8j1-16",
    hudsonOliveira = "select2-approver_user_id-result-87c0-17",
    barbaraPucci = "select2-approver_user_id-result-j5va-94",
    patriciaPeixoto = "select2-approver_user_id-result-5dsz-128",
    pedroUrban = "select2-approver_user_id-result-hc8e-129",
    marceloFontoura = "select2-approver_user_id-result-s0z0-130",
    clarissaPoletti = "select2-approver_user_id-result-6gk2-137",
    diretorgecom = "select2-approver_user_id-result-t4ce-193",
}