// cypress/DataParameters/Enums/showHideColumnsRequestReport.ts
export enum ShowHideColumnsRequestReport {
    type = 1,
    requestedOn = 2,
    assignmentDate = 3,
    serviceName = 4,
    hiredBy = 5,
    requester = 6,
    systemRequester = 7,
    status = 8,
    responsible = 9,
    costCenter = 10,
    supplier = 11,
    paymentMethod = 12,
    paymentCondition = 13,
    totalValue = 14
}