// cypress/DataParameters/Enums/paymentMethod.ts
export enum PaymentMethod {
    boleto = 'select2-payment-method-result-8om3-boleto',
    creditCard = 'select2-payment-method-result-67z7-cartao_credito',
    debitCard = 'select2-payment-method-result-wk3t-cartao_debito',
    cheque = 'select2-payment-method-result-u0uq-cheque',
    bankDeposit = 'select2-payment-method-result-laui-deposito_bancario',
    cash = 'select2-payment-method-result-ve2n-dinheiro',
    international = 'select2-payment-method-result-iakz-internacional',
    pix = 'select2-payment-method-result-qvua-pix',
}