// cypress/DataParameters/Enums/productCategory.ts
export enum ProductCategory {
    amostraGratisSolicitadaParaOFornecedor = 'select2-purchase_request_products0products0product_category_id-x5-result-tq3n-7',
    benfeitoriasEmImoveisDeTerceiros = 'select2-purchase_request_products0products0product_category_id-x5-result-hfm7-11',
    brindeMercadoriaDistribuidaGratuitamenteParaNossosClientesEQueNaoPodemosVenderExToalhaNecessaireEtc = 'select2-purchase_request_products0products0product_category_id-x5-result-muq9-5',
    computadoresEPerifericos = 'select2-purchase_request_products0products0product_category_id-x5-result-wd2q-16',
    epiLab = 'select2-purchase_request_products0products0product_category_id-x5-result-1bcd-17',
    epiManutencaoLimpeza = 'select2-purchase_request_products0products0product_category_id-x5-result-g8xg-21',
    instalacoes = 'select2-purchase_request_products0products0product_category_id-x5-result-nn5k-12',
    manutencaoPredial = 'select2-purchase_request_products0products0product_category_id-x5-result-gurw-20',
    maquinasEEquipamentos = 'select2-purchase_request_products0products0product_category_id-x5-result-q0ls-13',
    maquinasEEquipamentosNoLaboratorio = 'select2-purchase_request_products0products0product_category_id-x5-result-93gm-14',
    materialPromocionalMaterialPublicitarioGraficoQueVisaPromoverAPropagandaEMelhorVisualizacaoDosNossosProdutosExLaminasFoldersEtc = 'select2-purchase_request_products0products0product_category_id-x5-result-51u9-6',
    materialDeLimpezaECopa = 'select2-purchase_request_products0products0product_category_id-x5-result-th1j-3',
    materialDeUsoEConsumoLaboratorio = 'select2-purchase_request_products0products0product_category_id-x5-result-i4vo-1',
    materialDeEscritorio = 'select2-purchase_request_products0products0product_category_id-x5-result-mgn7-8',
    moveisEUtensilios = 'select2-purchase_request_products0products0product_category_id-x5-result-zw8w-15',
    obrasEmAndamento = 'select2-purchase_request_products0products0product_category_id-x5-result-9u51-10',
    outrosMateriaisDeUsoEConsumo = 'select2-purchase_request_products0products0product_category_id-x5-result-ovbz-19',
    outrosTiposDeProduto = 'select2-purchase_request_products0products0product_category_id-x5-result-7xdk-9',
    pecasReposicaoEmGeral = 'select2-purchase_request_products0products0product_category_id-x5-result-b7pb-22',
    pecasReposicaoEquipamentosEMaquinasLaboratorio = 'select2-purchase_request_products0products0product_category_id-x5-result-d2j5-23',
    revendaBonificadoMercadoriaQueVaiSerRevendidaETambemDistribuidaGratuitamenteParaOsNossosClientesBonificadosExViseiraGarrafaEtc = 'select2-purchase_request_products0products0product_category_id-x5-result-unid-4',
    uniformes = 'select2-purchase_request_products0products0product_category_id-x5-result-e6lp-18',
  }